# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server",
    "description": "",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/performances/{id}/lyrics": {
      "get": {
        "tags": [
          "performances"
        ],
        "operationId": "get_performance_lyrics",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lyrics for this performance. Returns performance-specific lyrics if set, otherwise falls back to the linked song's lyrics.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LyricsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "performances"
        ],
        "operationId": "put_performance_lyrics",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLyricsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Lyrics saved"
          },
          "404": {
            "description": "Performance not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "performances"
        ],
        "operationId": "delete_performance_lyrics",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Performance ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Performance lyrics override removed"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/discord/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "discord_callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Authorization code from Discord",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "CSRF state token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend, session cookie set"
          },
          "400": {
            "description": "Missing or invalid OAuth state"
          }
        }
      }
    },
...
    "/auth/twitch/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "twitch_callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Authorization code from Twitch",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "CSRF state token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend, session cookie set"
          },
          "400": {
            "description": "Missing or invalid OAuth state"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Body returned for all error responses.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "LyricsResponse": {
        "type": "object",
        "description": "Response body for lyrics GET endpoints.",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
      "UpdateLyricsRequest": {
        "type": "object",
        "description": "Request body for lyrics PUT endpoints.",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
    }
  }
}